+++ taf/updater/updater_pipeline.py
@@ -45,6 +45,56 @@ class RunMode(Enum):
 
 @define
 class UpdateState:
+    """
+    A class to manage the state of repository updates, encapsulating information about authentication and validation
+    of commits, and the management of repository data during updates.
+
+    Attributes:
+        auth_commits_since_last_validated (List[Any]): A list of fetched commits following the last validated commit
+        invalid_auth_commits (List[Any]): List of authenticated commits identified as invalid.
+        existing_repo (bool): Indicates if the repository already exists.
+        is_test_repo (bool): Flags the repository as a test repository. Is set based on the existance of a specific target file.
+        update_status (UpdateStatus): Update status - successful, partial, failed
+        update_successful (bool): Indicates whether the update was successful.
+        event (Event): event to be passed into the lifecycle handlers following the update.
+        users_auth_repo (Optional["AuthenticationRepository"]): User's authentication repository.
+        validation_auth_repo (Optional["AuthenticationRepository"]): Temporary authentication repository used for validation.
+        auth_repo_name (Optional[str]): Name of the authentication repository.
+        errors (Optional[List[Exception]]): Errors encountered during the update process.
+        warnings (Optional[List[Exception]]): Warnings issued during the update process.
+        targets_data (Dict[str, Any]): Data related to the target repositories, set based on the information in the
+            authentication repository.
+        last_validated_commit (str): The commit SHA of the last validated commit across the entire set of repositories.
+            This is only updated if the update is run without the --exclude-target option, so if the auth repo and all
+            target repositories were updated.
+        last_validated_data (Dict[str, Any]): Data set after each partial or successful update. If run with
+            --exclude-targets, last validated commit is set for each updated repo.
+        temp_target_repositories (Dict[str, "GitRepository"]): Temporary repositories created and cleaned up during
+            the update process.
+        users_target_repositories (Dict[str, "GitRepository"]): Permanent repositories within the user's local directory.
+        repos_on_disk (Dict[str, "GitRepository"]): Repositories that are already present in the user's local directory
+            at the start of the update.
+        repos_not_on_disk (Dict[str, "GitRepository"]): Repositories not present in the user's local directory at the
+            start of the update.
+        target_branches_data_from_auth_repo (Dict): Target repositories data, sorted by branches, based on information
+            from the authentication repository.
+        targets_data_by_auth_commits (Dict): Targets data organized by authentication commits.
+        old_heads_per_target_repos_branches (Dict[str, Dict[str, str]]): Old head commits per target repository branches.
+        fetched_commits_per_target_repos_branches (Dict[str, Dict[str, List[str]]]): Fetched commits per target repository branches.
+        validated_commits_per_target_repos_branches (Dict[str, Dict[str, str]]): Validated commits per target repository branches.
+        additional_commits_per_target_repos_branches (Dict[str, Dict[str, List[str]]]): Additional commits per target repository branches.
+        validated_auth_commits (List[str]): List of validated authenticated commits.
+        temp_root (TempPartition): Temporary storage partition used during updates.
+        update_handler (GitUpdater): update handler used to interface with TUF's updater.
+        clean_check_data (Dict[str, str]): A dictionary used while checking if the repositories are clean.
+        last_validated_data_per_repositories (Dict[str, Dict[str, str]]): Keeps track of last successfully validated
+            commits per branches of target repositories. Updated during the validation process.
+        all_targets_auth_commits (List[str]): All commits of the authentication repository that correspond
+            to commits of target repositories that need to be validated. If the previous update excluded some
+            target repositories, this list will not be the same as the list containing new auth repo commits.
+        is_partially_updated (bool): Indicates if the update was partial.
+    """
+
     auth_commits_since_last_validated: List[Any] = field(factory=list)
     invalid_auth_commits: List[Any] = field(factory=list)
     existing_repo: bool = field(default=False)
@@ -59,18 +109,11 @@ class UpdateState:
     warnings: Optional[List[Exception]] = field(default=None)
     targets_data: Dict[str, Any] = field(factory=dict)
     last_validated_commit: str = field(factory=str)
-    # repositories inside the temp folder which are created and cleaned up
-    # during the update process
+    last_validated_data: str = field(factory=dict)
     temp_target_repositories: Dict[str, "GitRepository"] = field(factory=dict)
-    # permanent repositories inside the user's local direcotry
     users_target_repositories: Dict[str, "GitRepository"] = field(factory=dict)
-    # a dictionary of repositories which are already present inside a user's local and
-    # directory when the update is started
     repos_on_disk: Dict[str, GitRepository] = field(factory=dict)
-    # a dictionary of repositories which are not present inside a user's local
-    # directory when the update is started
     repos_not_on_disk: Dict[str, GitRepository] = field(factory=dict)
-    # a list of repositoires that were added to repositories.json in one of new commits
     target_branches_data_from_auth_repo: Dict = field(factory=dict)
     targets_data_by_auth_commits: Dict = field(factory=dict)
     old_heads_per_target_repos_branches: Dict[str, Dict[str, str]] = field(factory=dict)
@@ -90,6 +133,8 @@ class UpdateState:
     last_validated_data_per_repositories: Dict[str, Dict[str, str]] = field(
         factory=dict
     )
+    all_targets_auth_commits: List[str] = field(factory=list)
+    is_partially_updated: bool = field(default=False)
 
 
 @attrs
@@ -170,7 +215,7 @@ class Pipeline:
                 if (
                     step_run_mode == RunMode.ALL or step_run_mode == self.run_mode
                 ) and (
-                    should_run_fn is None or should_run_fn()
+                    should_run_fn()
                 ):  # runs method like object
                     self.current_step = step
                     update_status = step()
@@ -221,22 +266,22 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
                 (
                     self.start_update,
                     RunMode.ALL,
-                    None,
+                    self.should_run_step_default,
                 ),
                 (
                     self.set_users_auth_repo,
                     RunMode.ALL,
-                    None,
+                    self.should_run_step_default,
                 ),
                 (
                     self.set_existing_target_repositories,
                     RunMode.UPDATE,
-                    None,
+                    self.should_run_step_default,
                 ),  # specify extra method here and runner will call to check if we can have it here
                 (
                     self.check_if_local_repositories_clean,
                     RunMode.UPDATE,
-                    None,
+                    self.should_run_step_default,
                 ),  # run only when want to updatae; valaidation doesn't change repop sstate; merge will fail without this
                 (
                     self.clone_auth_to_temp,
@@ -271,6 +316,16 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
                 # should_validate_target_repos
                 (
                     self.load_target_repositories,
+                    RunMode.ALL,
+                    self.should_run_step_default,
+                ),
+                (
+                    self.check_if_previous_update_partial,
+                    RunMode.ALL,
+                    self.should_run_step_default,
+                ),
+                (
+                    self.set_auth_commit_for_target_repos,
                     RunMode.ALL,
                     self.should_validate_target_repos,
                 ),
@@ -319,16 +374,20 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
                     RunMode.UPDATE,
                     self.should_validate_target_repos,
                 ),  # fetch commits; END UPDATE TARGET REPOS
-                (self.merge_commits, RunMode.UPDATE, None),  # merge fetched commits
+                (
+                    self.merge_commits,
+                    RunMode.UPDATE,
+                    self.should_validate_target_repos,
+                ),  # merge fetched commits
                 (
                     self.merge_auth_commits,
                     RunMode.UPDATE,
-                    None,
+                    self.should_run_step_default,
                 ),  # merge fetched commits
                 (
                     self.remove_temp_repositories,
                     RunMode.UPDATE,
-                    None,
+                    self.should_run_step_default,
                 ),  # only removes auth repo with --no-targets
                 (
                     self.set_target_repositories_data,
@@ -341,7 +400,7 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
                     self.should_validate_target_repos,
                 ),  # skipped with no-targets; prints all other commits that exist but are not merged
                 (self.check_pre_push_hook, RunMode.ALL, self.should_update_auth_repos),
-                (self.finish_update, RunMode.ALL, None),
+                (self.finish_update, RunMode.ALL, self.should_run_step_default),
             ],
             run_mode=(
                 RunMode.LOCAL_VALIDATION
@@ -378,6 +437,9 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
             )
         return self._output
 
+    def should_run_step_default(self):
+        return True
+
     def should_update_auth_repos(self):
         return True
 
@@ -387,9 +449,9 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
         if self.no_upstream:
             # check if self.state.event has changed.
             # if changed, validate the target repos
-            return (
-                self.state.event == Event.CHANGED or self.state.event == Event.PARTIAL
-            )
+            if self.state.event == Event.CHANGED or self.state.event == Event.PARTIAL:
+                return True
+            return self.state.is_partially_updated
         return True
 
     def start_update(self):
@@ -449,13 +511,28 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
                 repositoriesdb.clear_repositories_db()
         return UpdateStatus.SUCCESS
 
+    def _get_last_validated_commit(self, repo_name):
+        return self.state.last_validated_data.get(
+            repo_name, self.state.last_validated_commit
+        )
+
     def _set_last_validated_commit(self):
         if self.operation == OperationType.CLONE:
             settings.last_validated_commit[self.state.validation_auth_repo.name] = None
             self.state.last_validated_commit = None
+            self.state.last_validated_data = {}
         elif not self.validate_from_commit:
             users_auth_repo = AuthenticationRepository(path=self.auth_path)
-            last_validated_commit = users_auth_repo.last_validated_commit
+            # last_validated_commit is not updated after a partial update
+            # however, last_validated_data file is
+            # even thought we cannot claim that the system as a whole is valid
+            # after a partial update, there is no need to validate the auth repo
+            # from an older commit, just because one of more targets were skipped
+            self.state.last_validated_data = users_auth_repo.last_validated_data
+            last_validated_commit = self.state.last_validated_data.get(
+                users_auth_repo.name
+            )
+
             settings.last_validated_commit[
                 self.state.validation_auth_repo.name
             ] = last_validated_commit
@@ -468,7 +545,7 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
 
     def check_if_local_repositories_clean(self):
         try:
-            self.clean_check_data = {}
+            self.state.clean_check_data = {}
             # Early exit if the repository does not exist
             if not self.state.existing_repo:
                 return UpdateStatus.SUCCESS
@@ -541,6 +618,39 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
             self.state.errors.append(e)
             self.state.event = Event.FAILED
             return UpdateStatus.FAILED
+
+    def check_if_previous_update_partial(self):
+        """
+        Check if the previous update was a partial update
+        """
+
+        def _previous_update_partial():
+            if not self.state.users_auth_repo.last_validated_commit:
+                return True
+
+            # validate from last_validated_commit if last_validated_data does not exist
+            # (if last_validated_data was deleted)
+            if not self.state.users_auth_repo.last_validated_data:
+                return True
+            if (
+                self.state.users_auth_repo.last_validated_commit
+                != self.state.users_auth_repo.last_validated_data.get(
+                    self.state.users_auth_repo.name
+                )
+            ):
+                return True
+
+            last_validated_commits = set()
+            for repo_name in self.state.users_target_repositories:
+                if self.state.users_auth_repo.last_validated_data:
+                    if repo_name not in self.state.users_auth_repo.last_validated_data:
+                        return True
+                    last_validated_commits.add(
+                        self.state.users_auth_repo.last_validated_data[repo_name]
+                    )
+            return len(last_validated_commits) > 1
+
+        self.state.is_partially_updated = _previous_update_partial()
 
     def _check_if_target_repos_clean(self, target_repos, branches_per_repo):
         dirty_index_repos = []
@@ -949,6 +1059,40 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
             return UpdateStatus.FAILED
         return UpdateStatus.SUCCESS
 
+    def set_auth_commit_for_target_repos(self):
+
+        last_commits_per_repos = {
+            repo_name: self._get_last_validated_commit(repo_name)
+            for repo_name in self.state.users_target_repositories
+        }
+
+        last_validated_target_commits = list(set(last_commits_per_repos.values()))
+
+        if len(last_validated_target_commits) > 1:
+            # not all target repositories were updated at the same time
+            # updater was run with --exclude-targets
+            # check if the repositories are in sync according to that data
+            partially_validated_commits = (
+                self.state.users_auth_repo.auth_repo_commits_after_repos_last_validated(
+                    self.state.users_target_repositories.values(),
+                    self.state.last_validated_data,
+                )
+            )
+            all_auth_commits = partially_validated_commits
+            for commit in self.state.auth_commits_since_last_validated:
+                if commit not in all_auth_commits:
+                    all_auth_commits.append(commit)
+        else:
+            all_auth_commits = self.state.auth_commits_since_last_validated
+
+        self.state.all_targets_auth_commits = all_auth_commits
+
+        self.state.targets_data_by_auth_commits = (
+            self.state.users_auth_repo.targets_data_by_auth_commits(
+                all_auth_commits, last_commits_per_repos=last_commits_per_repos
+            )
+        )
+
     def load_target_repositories(self):
         taf_logger.debug(f"{self.state.auth_repo_name}: Loading target repositories...")
         try:
@@ -1045,12 +1189,12 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
         taf_logger.info(
             f"{self.state.auth_repo_name}: Validating initial state of target repositories..."
         )
+        # if update was run with --exclude-target options, some target repositories might have been
+        # validated up to an older commit, or never validated at all
+        # that means that different repositories need to be validated from different start commits
+
         try:
-            self.state.targets_data_by_auth_commits = (
-                self.state.users_auth_repo.targets_data_by_auth_commits(
-                    self.state.auth_commits_since_last_validated
-                )
-            )
+
             self.state.old_heads_per_target_repos_branches = defaultdict(dict)
             is_initial_state_in_sync = True
             # if last validated commit was not manually modified (set to a newer commit)
@@ -1068,17 +1212,25 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
                             not in self.state.targets_data_by_auth_commits
                         ):
                             continue
+
                         self.state.old_heads_per_target_repos_branches[
                             repository.name
                         ] = {}
+                        repo_last_validated_commit = self._get_last_validated_commit(
+                            repository.name
+                        )
                         last_validated_repository_commits_data = (
                             self.state.targets_data_by_auth_commits[
                                 repository.name
-                            ].get(self.state.last_validated_commit, {})
+                            ].get(repo_last_validated_commit, {})
                         )
 
                         if last_validated_repository_commits_data:
-                            if repository.name in self.state.repos_not_on_disk:
+                            if (
+                                repository.name in self.state.repos_not_on_disk
+                                and self._get_last_validated_commit(repository.name)
+                                is not None
+                            ):
                                 is_initial_state_in_sync = False
                                 break
                             if not self._is_repository_in_sync(
@@ -1162,10 +1314,14 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
 
     def _update_state_for_initial_sync(self):
         self.state.last_validated_commit = None
+        self.state.last_validated_data = {}
         self.state.auth_commits_since_last_validated = (
             self.state.users_auth_repo.all_commits_on_branch(
                 self.state.users_auth_repo.default_branch
             )
+        )
+        self.state.all_targets_auth_commits = (
+            self.state.auth_commits_since_last_validated
         )
         # append fetched commits
         if self.state.update_handler is not None and self.state.update_handler.commits:
@@ -1174,7 +1330,7 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
             )
         self.state.targets_data_by_auth_commits = (
             self.state.users_auth_repo.targets_data_by_auth_commits(
-                self.state.auth_commits_since_last_validated
+                self.state.all_targets_auth_commits
             )
         )
 
@@ -1320,10 +1476,9 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
             # need to be set to old head since that is the last validated target
             self.state.validated_commits_per_target_repos_branches = defaultdict(dict)
 
-            self.last_validated_data_per_repositories = defaultdict(dict)
+            self.state.last_validated_data_per_repositories = defaultdict(dict)
             self.state.validated_auth_commits = []
-
-            for auth_commit in self.state.auth_commits_since_last_validated:
+            for auth_commit in self.state.all_targets_auth_commits:
                 for repository in self.state.temp_target_repositories.values():
                     if repository.name not in self.state.targets_data_by_auth_commits:
                         continue
@@ -1341,24 +1496,31 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
                     )
                     current_commit = current_targets_data["commit"]
                     if not len(
-                        self.last_validated_data_per_repositories[repository.name]
+                        self.state.last_validated_data_per_repositories[repository.name]
                     ):
+                        last_validated_target_auth_commit = (
+                            self._get_last_validated_commit(repository.name)
+                        )
                         current_head_commit_and_branch = (
                             self.state.targets_data_by_auth_commits[
                                 repository.name
-                            ].get(self.state.last_validated_commit, {})
+                            ].get(last_validated_target_auth_commit, {})
                         )
                         previous_branch = current_head_commit_and_branch.get("branch")
                         previous_commit = current_head_commit_and_branch.get("commit")
                         if previous_commit is not None and previous_branch is None:
                             previous_branch = repository.default_branch
                     else:
-                        previous_branch = self.last_validated_data_per_repositories[
-                            repository.name
-                        ].get("branch")
-                        previous_commit = self.last_validated_data_per_repositories[
-                            repository.name
-                        ]["commit"]
+                        previous_branch = (
+                            self.state.last_validated_data_per_repositories[
+                                repository.name
+                            ].get("branch")
+                        )
+                        previous_commit = (
+                            self.state.last_validated_data_per_repositories[
+                                repository.name
+                            ]["commit"]
+                        )
 
                     target_commits_from_target_repo = (
                         self.state.fetched_commits_per_target_repos_branches[
@@ -1376,7 +1538,7 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
                         auth_commit,
                     )
 
-                    self.last_validated_data_per_repositories[repository.name] = {
+                    self.state.last_validated_data_per_repositories[repository.name] = {
                         "commit": validated_commit,
                         "branch": current_branch,
                     }
@@ -1602,7 +1764,7 @@ but commit not on branch {current_branch}"
                 return self.state.update_status
             for repository in self.state.users_target_repositories.values():
                 # this will only include branches that were, at least partially, validated (up until a certain point)
-                last_branch = self.last_validated_data_per_repositories[
+                last_branch = self.state.last_validated_data_per_repositories[
                     repository.name
                 ]["branch"]
                 for (
@@ -1638,7 +1800,7 @@ but commit not on branch {current_branch}"
             f"{self.state.auth_repo_name}: Merging commit into auth repo..."
         )
         try:
-            if self.only_validate or self.excluded_target_globs:
+            if self.only_validate:
                 return self.state.update_status
             if self.state.update_status == UpdateStatus.FAILED:
                 return self.state.update_status
@@ -1658,8 +1820,17 @@ but commit not on branch {current_branch}"
                 last_commit,
                 True,
             )
-            # update the last validated commit
-            self.state.users_auth_repo.set_last_validated_commit(last_commit)
+
+            # store information about which target (data) repositories were updated
+            # some might have been omitted if the update was run with --exclude-target
+            last_validated_data = self.state.last_validated_data or {}
+            for repo in self.state.users_target_repositories.keys():
+                last_validated_data[repo] = last_commit
+            last_validated_data[self.state.users_auth_repo.name] = last_commit
+            self.state.users_auth_repo.set_last_validated_data(
+                last_validated_data,
+                set_last_validated_commit=not bool(self.excluded_target_globs),
+            )
 
             return self.state.update_status
         except Exception as e: