+++ CHANGELOG.md
@@ -9,9 +9,15 @@ and this project adheres to [Semantic Versioning][semver].
 
 ### Added
 
+- Implement full partial update. Store last validated commit per repo ([559)])
+
 ### Changed
 
 ### Fixed
+
+
+[559]: https://github.com/openlawlibrary/taf/pull/558
+
 
 ## [0.32.4]